Added class SingleWiFiScanWindows
This class has a single method that calls the netsh command, combines the shell
output lines into a single string, and returns that string. Plan to add a second
method that is nearly identical, but it will accept a string to be used as the
network interface device, allowing the user to specify which WiFi device to use
for the scan should there be more than one present.