Broke up the previous scan to accommodate overloading.
The previous scan method was separated into two methods; one that accepts a process
builder object (runScan) and another that accepts a string (scan). The string
identifies the WiFi interface to be used, which in turn is included in the process
builder object passed into runScan. If no string is provided, the method is
overloaded such that it will pass a different process builder object into runScan
that omits the specified interface.