Added class ParseGPSData
In progress, no content yet. Conducting Domain research to better understand
the content of GPS NMEA sentences. This will shape the structure of the Maps
for each unique NMEA sentence. The class will return a List of JSONObjects
that can be merged (via another class) with the JSONObjects produced by the
ParseWiFiDataWindows class.