Completed the initial version of ParseGPSData class
This class consists of a Map of data labels, 3 String arrays, a constructor, one
method and a main() for testing. The constructor combines the 3 String arrays to
initialize the Map; this is done in the constructor. The three string arrays are
in a particular order to enable proper label matching to data reported by the GPS
device being used, based on standard GPS NMEA sentences. The method takes in a
string of NMEA GPS data sentences and parses the contained data into a Map of GPS
data. Duplicate keys result in overwriting existing values, this occurs 6 times
(utcTime, latitude, northSouth, longitude, eastWest, hdop). The method returns a
JSON object; the intent is to merge this JSON object with the JSON object
returned by ParseWiFiDataWindows, resulting in the final JSON Object to be stored
in the database.