Added new method getGPSDataIgnoreGPGSV
Added a new method, getGPSDataIgnoreGPGSV, and its overloaded method. This method
ignores collecting GPGSV NMEA sentences as the info provided is likely not very
useful to NAn. Due to the way the GPS devices collects and passes NMEA sentences
to the program, it takes longer waiting for all 4 NMEA sentences to come through
rather than just waiting for the 3 NMEA sentences of value (GPGGA, GPGSA, GPRMC).
It usually takes less than 0.9 seconds to get the GPGGA, GPGSA, GPRMC sentences
whereas it might take an additonal 2 to 3 seconds to also get the GPGSV sentence.
The purpose of this method is to avoid the additional 2 or 3 second delay when the
GPGSV NMEA sentence is likely not very useful for our application. Might cut GPGGA
sentences in the future, as the only additional data provided is the altitude and
a few other items that are less significant.