New class MergeJSONData
MergeJSONData take a list of WiFi JSON objects and a single GPS JSON object and
adds all key-value pairs from the GPS JSON object to each WiFi JSON object. It
then returns a list of those merged JSON objects; each JSON object in the list is
a stand-alone entry into the database. There is redundancy/wasted space in a
sense; the GPS data is the same for each unique MAC that was picked up in the WiFi
scan. However, including the GPS in this way makes it possible to take each
individual MAC as a single data point, independent of other data. Perhaps we
optimize this in the future, it'll probably depend on the database, but the saved
memory is likely not worth introducing additional dependencies and points of
failure. The main() method is included for testing purposes only.