Added class to create a .kml from the list of JSON Objects
First draft of a class to create a .kml file from the List of
JSON objects produced from the MergeJSONData class. Currently
missing a few field items that need to be added to the description
of the Placemark, namely "MAC". Will probably remove the "Course
Over Ground" and "Speed Over Ground", as they are a bit misleading
when satellite fix isn't very good, or the scan is done indoors.