Added logic to handle missing lat/long.
In the event that gps data is collect but the geospatial
information (latitude and or longitude) is
lost/corrupt/missing, the new logic ensures data points
are not added to the .kml file, as the .kml file needs
that geospatial data to plot the placemark. The .kml file
will not be created if no placemarks have geospatial data
but will be created if at least one placemark does.

Also updated the file name fo the created .kml to lead
with the date and time of data collection.